python/requirements.txt
flask
keras==2.2.4
tensorflow==1.12
flask-socketio
msgpack-numpy
socketIO_client
pymongo
Gevent
psycopg2